requirements.txt
Flask
gunicorn
stripe
requests
Pillow
qrcode